requirements.txt
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-joblib-launcher==1.2.0
hydra-submitit-launcher==1.2.0
matplotlib==3.8.1
numpy==1.26.1
omegaconf==2.3.0
pandas==2.0.3
Pillow==10.1.0
POT==0.9.1
pytorch_lightning==1.9.5
PyYAML==6.0.2
rich==13.7.1
scikit_learn==1.3.2
torch_geometric==2.5.3
torch-cluster==1.6.3
torchdyn==1.0.6
tqdm==4.66.1
wandb==0.16.0
python-dotenv==1.0.1